fix: Do not pass unexpected arguments with callPackage
We are overriding `callPackage` to allow replacing `pcre2` argument for compatibility with older PHP versions.
The override causes the argument to be passed to packages unconditionally.

It was not an issue previously since, until recently, `generic.nix` was only calling `php-packages.nix`.

This changed with <https://github.com/NixOS/nixpkgs/commit/c038b79502f379b1fe988520fc0d5f01c5799156>, which introduced an `install-pear-nozlib.phar` package, also instantiated with `callPackage`, that takes only `fetchurl` as an argument.

As a result, PHP from Nixpkgs after this commit would not evaluate:

    error: function 'anonymous lambda' called with unexpected argument 'pkgs'

Let’s only pass the arguments when they are expected.

Fixes: https://github.com/fossar/nix-phps/issues/310