php{7[1234],80}.extension.gettext: Fix/skip failing test on Darwin with gettext 0.22.5
The incoming bump of the gettext library to 0.22.5 reportedly [1] introduces an assertion that `category` argument of `dcgettext` and `dcngettext` functions is not `LC_ALL`, causing PHP crash with SIGABRT when that happens.  This should be rare in practice – I only found one example of a decade-old WordPress plug-in doing that [2].  However, one test was passing `0` as the category argument to those functions, which triggered the assertion on Darwin because it defines `LC_ALL` as 0. On Linux, glibc defines `LC_CTYPE` as 0 so it is not problem there.

PHP 8.2 introduced a validation step [1] that will make the functions throw on receiving `LC_ALL`, and we cherry-picked it to 8.0.
We decided just to remove the offending test on older versions since the crash is unlikely in practice and backporting the fix is too much bother.

[1]: https://github.com/php/php-src/commit/9999a0cb757344974889a6f548727de6f2c3c10d
[2]: https://github.com/search?utf8=%E2%9C%93&q=%2Fdcn%3Fgettext.%2B%28LC_ALL%7C0%29%2F+language%3APHP&type=code